refactor: Remove unused code and styles related to the map component
