+++ package.json
@@ -12,7 +12,6 @@
     "@mui/icons-material": "^5.14.14",
     "@mui/material": "^5.14.14",
     "@types/date-fns": "^2.6.0",
-    "axios": "^1.6.8",
     "date-fns": "^2.30.0",
     "mapbox-gl": "^2.15.0",
     "next": "13.5.4",
