+++ tsconfig.json
@@ -2,7 +2,7 @@
     "compilerOptions": {
       "target": "es5",
       "lib": ["dom", "dom.iterable", "esnext"],
-      "typeRoots": ["./types", "node_modules/@types"],
+      "typeRoots": ["src/app/types", "node_modules/@types"],
       "allowJs": true,
       "skipLibCheck": true,
       "strict": true,
